Add function to display information on OLED screen, add function to calculate battery percentage.

+++ Basic_TX_ESP32/Basic_TX_ESP32.ino
@@ -59,7 +59,7 @@ typedef enum {
 struct __attribute__((packed)) RadioPacket {
   Opcode_t opcode;
   uint8_t fromID;
-  int pin;
+  uint16_t pin;
   uint8_t controlTime,
     periodTime,
     pauseTime;
